requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
openai==1.51.2
Flask-Cors==5.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
